Move createFulfilledPromise to internal
